+++ module/package.json
